pyproject.toml
[tool.poetry]
name = "humming_bot_graph"
version = "0.1.0"
description = ""
authors = ["Yanis"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.2.2"
jupyter = "^1.0.0"
python-binance = "^0.7.9"
plotly = "^4.14.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
